tests/knowledge/rules/aws/non_context_aware/test_ensure_ecs_cluster_enable_container_insights_rule.py
import unittest

from cloudrail.knowledge.context.aws.ecs.ecs_cluster import EcsCluster
from cloudrail.knowledge.context.environment_context import EnvironmentContext
from cloudrail.knowledge.rules.aws.non_context_aware.ensure_ecs_cluster_enable_container_insights_rule import \
    EnsureEcsClusterEnableContainerInsightsRule
from cloudrail.knowledge.rules.base_rule import RuleResultType
from cloudrail.dev_tools.rule_test_utils import create_empty_entity


class TestEnsureEcsClusterEnableContainerInsightsRule(unittest.TestCase):
    def setUp(self):
        self.rule = EnsureEcsClusterEnableContainerInsightsRule()

    def test_non_car_ecs_cluster_container_insights_enabled_fail(self):
        # Arrange
        ecs_cluster: EcsCluster = create_empty_entity(EcsCluster)
        ecs_cluster.is_container_insights_enabled = False
        context = EnvironmentContext(ecs_cluster_list=[ecs_cluster])
        # Act
        result = self.rule.run(context, {})
        # Assert
        self.assertEqual(RuleResultType.FAILED, result.status)
        self.assertEqual(1, len(result.issues))

    def test_non_car_ecs_cluster_container_insights_enabled_pass(self):
        # Arrange
        ecs_cluster: EcsCluster = create_empty_entity(EcsCluster)
        ecs_cluster.is_container_insights_enabled = True
        context = EnvironmentContext(ecs_cluster_list=[ecs_cluster])
        # Act
        result = self.rule.run(context, {})
        # Assert
        self.assertEqual(RuleResultType.SUCCESS, result.status)
        self.assertEqual(0, len(result.issues))
